+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/controllers/AppExceptionHandler.java
@@ -2,6 +2,7 @@ package com.wojciechbarwinski.demo.epic_board_games_shop.controllers;
 
 
 import com.wojciechbarwinski.demo.epic_board_games_shop.exceptions.ErrorResponse;
+import com.wojciechbarwinski.demo.epic_board_games_shop.exceptions.MappingToDTOException;
 import com.wojciechbarwinski.demo.epic_board_games_shop.exceptions.MissingCredentialsException;
 import com.wojciechbarwinski.demo.epic_board_games_shop.exceptions.ProductsNotFoundException;
 import com.wojciechbarwinski.demo.epic_board_games_shop.security.exceptions.ApplicationSecurityException;
@@ -33,6 +34,13 @@ public class AppExceptionHandler {
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ExceptionHandler(MissingCredentialsException.class)
     public ErrorResponse<String> missingCredentialsException(MissingCredentialsException exception) {
+
+        return new ErrorResponse<>(exception.getMessage());
+    }
+
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ExceptionHandler(MappingToDTOException.class)
+    public ErrorResponse<String> mappingToDTOException(MappingToDTOException exception) {
 
         return new ErrorResponse<>(exception.getMessage());
     }
