+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/dtos/LoginDTO.java
@@ -1,5 +1,4 @@
 package com.wojciechbarwinski.demo.epic_board_games_shop.dtos;
 
 public record LoginDTO(String username, String password) {
-
 }
