+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/dtos/OrderResponseDTO.java
@@ -22,5 +22,7 @@ public class OrderResponseDTO {
 
     private AddressDTO addressToSend;
 
+    private String status;
+
     private List<OrderLineDTO> orderLineDTOs;
 }
