+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/entities/OrderStatus.java
@@ -3,6 +3,7 @@ package com.wojciechbarwinski.demo.epic_board_games_shop.entities;
 public enum OrderStatus {
 
     PLACED,
+    CONFIRMED,
     PAYMENT_VERIFICATION,
     PAID,
     RECEIVED_BY_WAREHOUSE,
