+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/exceptions/MappingToDTOException.java
@@ -1,0 +1,8 @@
+package com.wojciechbarwinski.demo.epic_board_games_shop.exceptions;
+
+public class MappingToDTOException extends ApplicationException{
+
+    public MappingToDTOException(String message) {
+        super(message);
+    }
+}
